+++ core-bundle/src/Resources/contao/modules/ModulePersonalData.php
@@ -199,6 +199,8 @@ class ModulePersonalData extends Module
 			/** @var Widget $objWidget */
 			$objWidget = new $strClass($strClass::getAttributesFromDca($arrData, $field, $varValue, $field, $strTable, $this));
 
+			// Append the module ID to prevent duplicate IDs (see #1493)
+			$objWidget->id .= '_' . $this->id;
 			$objWidget->storeValues = true;
 			$objWidget->rowClass = 'row_' . $row . (($row == 0) ? ' row_first' : '') . ((($row % 2) == 0) ? ' even' : ' odd');