+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+ * Add the page picker to the meta fields in file manager (see #1568).
  * Remove the "flash movie" front end module.
  * Add extended video support (see #1348).
  * Pass all search result data to the search template (see #1558).