+++ core-bundle/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## DEV
+
+ * Make the session listener compatible with Symfony 3.4.12.
+
 ## 4.4.19 (2018-06-18)
 
  * Show the 404 page if there is no match in the set of potential pages (see #1522).